+++ packages/semantic-release/src/index.js
@@ -14,6 +14,12 @@ module.exports = {
     "@semantic-release/commit-analyzer",
     "@semantic-release/release-notes-generator",
     "@semantic-release/changelog",
+    [
+      "@semantic-release/exec",
+      {
+        prepareCmd: "yarn run fix",
+      },
+    ],
     "@semantic-release/npm",
     "@semantic-release/github",
     "@semantic-release/git",